backend/requirements.txt
Django==4.2.27
djangorestframework==3.15.2
django-cors-headers==4.4.0
requests==2.32.3
python-dotenv==1.0.0
gunicorn==23.0.0
dj-database-url==2.1.0
openpyxl==3.1.5
psycopg2-binary==2.9.9
